Making sure the analysis board link works also for PGNs with starting FENs (#139)
Lichess.org has two ways of launching an analysis board. One is through a FEN and the other
is through a PGN and a color. The benefit of the PGN version is that it allows the user to
step back through the move line and evaluate the position at any of the played moves. If the
repertoire doesn't start on move 1, ie it has a FEN code as a starting positionin the PGN
file, then using the PGN url doesn't work. So to utilize the possibility of the GN url when
we can, we use that url when the repoertoire starts at move 1, or else the FEN url.